% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quandl_unemployement_data.R
\name{read_quandl_unemployement_data}
\alias{read_quandl_unemployement_data}
\title{Read Unemployement US data}
\usage{
read_quandl_unemployement_data()
}
\value{
data.frame
}
\description{
Read Unemployement US data
}
\examples{
\donotrun{read_quandl_unemployement_data()}
}
