% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_operations.R
\name{rest_get_yearly_unemployment_data}
\alias{rest_get_yearly_unemployment_data}
\title{Main Operation of GET yearly unemployment data REST service}
\usage{
rest_get_yearly_unemployment_data(from, to)
}
\description{
Main Operation of GET yearly unemployment data REST service
}
