% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_operations.R
\name{rest_store_yearly_unemployment_data}
\alias{rest_store_yearly_unemployment_data}
\title{Main Operation of POST storing yearly unemployment data REST service}
\usage{
rest_store_yearly_unemployment_data()
}
\description{
Main Operation of POST storing yearly unemployment data REST service
}
