% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_api.R
\name{run_api}
\alias{run_api}
\title{Run the housing prediction API}
\usage{
run_api(host = "0.0.0.0", port = 8000, ...)
}
\arguments{
\item{host}{Character. Address to run API on}

\item{port}{Integer. Port to run API on}

\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run the housing prediction API
}
\examples{
\dontrun{
run_api()
run_api(swagger = TRUE, port = 8000)
}
}
